Modified number plate string creation algorithm
Many a time, there are detections of two characters together, etc. This modification makes sure that there are no nested detections. The algorithm is as follows. First, we find a (kind of) mode of all detection widths. Then we sort all the detections in decreasing order of areas. As we pick each detection, we check if the current box lies inside an already picked box. If the current box does not lie inside any already chosen box, then we chose the current box. If the current box does lie inside a picked box, we choose one of the two boxes - the one whose width is closer to the modeWidth calculated.

To find the modeWidth, we put all the boxes into different bins and create a histogram. If more than half the boxes fall into a single bin, we pick the average of box widths in that bin as modewidth. We start with a bin window size of 5 and increment it till full image width / 4. We return at the window size where the above condition is satisfied. If the above condition is not satisfied for any window size, we return without creating the string. Finally we add only those detections whose width is less than double the mode width.